apps/agent/pyproject.toml
[project]
name = "tripgenius-agent"
version = "0.1.0"
description = "TripGenius Travel Planning Agent"
requires-python = ">=3.12"
dependencies = [
  "langchain==1.2.0",
  "langgraph==1.0.5",
  "langsmith>=0.4.49",
  "openai>=1.68.2,<2.0.0",
  "fastapi>=0.115.5,<1.0.0",
  "uvicorn>=0.29.0,<1.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "langgraph-cli[inmem]>=0.4.11",
  "langchain-openai>=1.1.0",
  "copilotkit==0.1.86",
  "langgraph-api>=0.7.70",
  "pip>=26.0.1",
  "langchain-anthropic>=1.3.4",
  "requests>=2.31.0",
]

[tool.uv]
index-url = "https://pypi.org/simple"
